+++ rules/import.js
@@ -16,5 +16,8 @@ module.exports = {
    */
   'import/order': ['warn', {
     groups: ['builtin', 'external', 'internal', 'parent', 'sibling', 'index'],
+    alphabetize: {
+      order: 'asc',
+    },
   }],
 };